evaluaciones/evaluacion3/gestionAPP/requirements.txt
django==5.0
django-ninja==1.0.1
pyjwt==2.8.0